# .api/apis/cvcrm/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "Ambiente de Integração",
      "url": "https://{dominiodocliente}.cvcrm.com.br/api",
      "variables": {
        "dominiodocliente": {
          "default": "integracao"
        }
      }
    },
    {
      "description": "Ambiente de Desenvolvimento",
      "url": "https://dev.cvcrm.com.br/api"
    }
  ],
  "info": {
    "description": "Esta API é responsável por cadastrar, alterar e exibir dados dos empreendimentos.",
    "version": "1.0.0",
    "title": "API de Empreendimentos"
  },
...
  "paths": {
...
    "/v1/cv/unidade/{idempreendimento}/{id}": {
      "get": {
        "summary": "Retorna unidade.",
        "tags": [
          "Unidade"
        ],
        "description": "Esta Api retorna os dados de uma Unidade",
        "parameters": [
          {
            "in": "header",
            "name": "email",
            "schema": {
              "example": "email@email.com",
              "type": "string"
            },
            "required": true,
            "description": "E-mail de autenticação para retorno de unidade."
          },
          {
            "in": "header",
            "name": "token",
            "schema": {
              "example": "79d965e5027b7423423f2471824e799949c75f9ce",
              "type": "string"
            },
            "required": true,
            "description": "Token de autenticação para retorno de unidade."
          },
          {
            "name": "idempreendimento",
            "in": "path",
            "description": "id do empreendimento da unidade",
            "required": true,
            "schema": {
              "example": 520,
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id da unidade ou código interno ",
            "required": true,
            "schema": {
              "example": 1308,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "nome",
            "description": "Nome da unidade",
            "schema": {
              "example": "unidade X",
              "type": "string"
            }
          }
        ],
        "operationId": "retornarEmpreendimentoPorUnidade",
        "responses": {
          "200": {
            "description": "Unidade",
            "content": {
              "application/json": {
                "schema": {
                  "description": "JSON de retorno das unidades.",
                  "type": "object",
                  "properties": {
                    "Pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "example": 1,
                          "type": "integer"
                        },
                        "perPage": {
                          "example": 3,
                          "type": "integer"
                        },
                        "currentPage": {
                          "example": 1,
                          "type": "integer"
                        },
                        "lastPage": {
                          "example": 1,
                          "type": "integer"
                        }
                      }
                    },
                    "data": {
                      "description": "JSON de retorno das unidades novas.",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "idempreendimento": {
                            "example": 1186,
                            "type": "integer"
                          },
                          "idempreendimento_int": {
                            "example": "1186",
                            "type": "string"
                          },
                          "nome_empreendimento": {
                            "example": "Empreendimento X",
                            "type": "string"
                          },
                          "idetapa": {
                            "example": 1186,
                            "type": "integer"
                          },
                          "idetapa_int": {
                            "example": 1186,
                            "type": "integer"
                          },
                          "etapa": {
                            "example": "1ª ETAPA",
                            "type": "string"
                          },
                          "idbloco": {
                            "example": 1186,
                            "type": "integer"
                          },
                          "idbloco_int": {
                            "example": 1186,
                            "type": "integer"
                          },
                          "bloco": {
                            "example": "BLOCO 01",
                            "type": "string"
                          },
                          "nome": {
                            "example": "Unidade 001",
                            "type": "string"
                          },
                          "area_privativa": {
                            "example": "122.00",
                            "type": "string"
                          },
                          "idunidade": {
                            "example": "12345",
                            "type": "string"
                          },
                          "idunidade_int": {
                            "example": "123",
                            "type": "string"
                          },
                          "vagas_garagem": {
                            "example": "1",
                            "type": "string"
                          },
                          "andar": {
                            "example": "7",
                            "type": "string"
                          },
                          "area_comum": {
                            "example": "33.39",
                            "type": "string"
                          },
                          "coluna": {
                            "example": 10,
                            "type": "integer"
                          },
                          "posicao": {
                            "example": "Leste",
                            "type": "string"
                          },
                          "tipologia": {
                            "example": "3 suites com terraço gourmet",
                            "type": "string"
                          },
                          "tipo": {
                            "example": "Apartamento",
                            "type": "string"
                          },
                          "idtipo_int": {
                            "example": "123",
                            "type": "string"
                          },
                          "empresa_terceirizacao": {
                            "example": "Nome da empresa de terceirização",
                            "type": "string"
                          },
                          "valor": {
                            "example": "9999.99",
                            "type": "string"
                          },
                          "valor_avaliacao": {
                            "example": "497242.40",
                            "type": "string"
                          },
                          "data_bloqueio": {
                            "example": "01/10/2024",
                            "type": "string"
                          },
                          "data_entrega": {
                            "example": "01/01/2023",
                            "type": "string"
                          },
                          "data_entrega_chaves": {
                            "example": "01/01/2023",
                            "type": "string"
                          },
                          "agendar_a_partir": {
                            "example": "01/01/2023",
                            "type": "string"
                          },
                          "liberar_a_partir": {
                            "example": "01/01/2023",
                            "type": "string"
                          },
                          "plantas": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "example": {
                                "idplanta": 1,
                                "nome": "planta 1"
                              }
                            },
                            "example": [
                              {
                                "idplanta": 1,
                                "nome": "planta 1"
                              },
                              {
                                "idplanta": 2,
                                "nome": "planta 2"
                              }
                            ]
                          },
                          "campos_adicionais": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "nome_referencia": {
                                  "type": "string",
                                  "example": "adc_valor_adicional",
                                  "description": "Nome de referência do campo adicional de unidades novas."
                                },
                                "valor": {
                                  "type": "string",
                                  "example": "1.000,00",
                                  "description": "Valor do campo adicional em unidades novas."
                                }
                              },
                              "example": {
                                "nome_referencia": "adc_valor_adicional",
                                "valor": "1.000,00"
                              }
                            },
                            "example": [
                              {
                                "nome_referencia": "adc_valor_adicional",
                                "valor": "1.000,00"
                              }
                            ]
                          },
                          "situacao": {
                            "description": "Esquema de retorno da situação da unidade.",
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "vendida": {
                                  "example": "1332",
                                  "type": "string"
                                },
                                "vendida_idsituacao": {
                                  "example": "3",
                                  "type": "string"
                                },
                                "idmotivo": {
                                  "example": "1",
                                  "type": "string"
                                },
                                "tipo": {
                                  "description": "Informações do tipo.",
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "example": "1",
                                      "type": "string"
                                    },
                                    "nome": {
                                      "example": "Garagem",
                                      "type": "string"
                                    }
                                  },
                                  "example": {
                                    "id": "1",
                                    "nome": "Garagem"
                                  }
                                },
                                "situacao_para_venda": {
                                  "example": 1,
                                  "type": "integer"
                                },
                                "situacao_mapa_disponibilidade": {
                                  "example": 1,
                                  "type": "integer",
                                  "description": "Situação da unidade no mapa de disponibilidade (1 - Disponível; 2 - Reservada; 3 - Vendida; 4 - Bloqueada; 5 - Em processo de reserva)"
                                },
                                "possui_reserva_solicitacao_distrato": {
                                  "example": "N",
                                  "type": "string"
                                }
                              },
                              "example": {
                                "vendida": "1332",
                                "vendida_idsituacao": "3",
                                "idmotivo": "1",
                                "situacao_para_venda": 1,
                                "situacao_mapa_disponibilidade": 1,
                                "possui_reserva_solicitacao_distrato": "N"
                              }
                            },
                            "example": [
                              {
                                "vendida": "1332",
                                "vendida_idsituacao": "3",
                                "idmotivo": "1",
                                "situacao_para_venda": 1,
                                "situacao_mapa_disponibilidade": 1,
                                "possui_reserva_solicitacao_distrato": "N"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "CadastroUnidade": {
                    "summary": "Unidade 001"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro ao tentar retornar unidade.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao tentar retornar unidade."
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    }
                  }
                },
                "examples": {
                  "msgErro400": {
                    "summary": "Erro 400"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}